backend/requirements.txt
affine==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aioredis==2.0.1
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
boto3==1.35.33
botocore==1.35.33
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.0.0
flake8-docstrings==1.7.0
flower==2.0.1
fonttools==4.54.1
frozenlist==1.4.1
humanize==4.10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
kombu==5.4.2
matplotlib==3.9.2
mccabe==0.7.0
multidict==6.1.0
numpy==2.0.2
packaging==24.1
pillow==10.4.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
PyJWT==2.9.0
pyparsing==3.1.4
pyproj==3.6.1
pystac==1.10.1
pystac-client==0.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
rasterio==1.4.1
redis==5.1.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
s3transfer==0.10.2
sat-search==0.1.0
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
yarl==1.13.1
zipp==3.20.2
